+++ core/site.ts
@@ -125,8 +125,8 @@ export default class Site {
   /** Pages created with site.page() */
   scopedPages = new Map<string, Data[]>();
 
-  /** Global components shared by all templates */
-  globalComponents: Components = new Map();
+  /** Components created with site.component() */
+  scopedComponents = new Map<string, Components>();
 
   /** Hooks installed by the plugins */
   // deno-lint-ignore no-explicit-any
@@ -163,6 +163,7 @@ export default class Site {
       components,
       scopedData: this.scopedData,
       scopedPages: this.scopedPages,
+      scopedComponents: this.scopedComponents,
       prettyUrls,
     });
 
@@ -422,9 +423,11 @@ export default class Site {
   }
 
   /** Register an extra component accesible by the layouts */
-  component(context: string, component: Component): this {
+  component(context: string, component: Component, scope = "/"): this {
     const pieces = context.split(".");
-    let components = this.globalComponents;
+    const scopedComponents: Components = this.scopedComponents.get(scope) ||
+      new Map();
+    let components: Components = scopedComponents;
 
     while (pieces.length) {
       const name = pieces.shift()!;
@@ -435,6 +438,7 @@ export default class Site {
     }
 
     components.set(component.name, component);
+    this.scopedComponents.set(scope, scopedComponents);
     return this;
   }
 
@@ -529,10 +533,7 @@ export default class Site {
     // Get the site content
     const isDev = Deno.env.get("LUME_ENV") === "development";
     const [_pages, _staticFiles] = await this.source.build(
-      this.globalComponents,
-      [
-        (_, page) => !page?.data.draft || isDev,
-      ],
+      (_, page) => !page?.data.draft || isDev,
     );
 
     // Save static files into site.files
@@ -582,11 +583,8 @@ export default class Site {
     // Get the site content
     const isDev = Deno.env.get("LUME_ENV") === "development";
     const [_pages, _staticFiles] = await this.source.build(
-      this.globalComponents,
-      [
-        (_, page) => !page?.data.draft || isDev,
-        this.scopes.getFilter(files),
-      ],
+      (_, page) => !page?.data.draft || isDev,
+      this.scopes.getFilter(files),
     );
 
     // Build the pages and save static files into site.files
@@ -688,12 +686,9 @@ export default class Site {
 
     // Get the site content
     const [pages] = await this.source.build(
-      this.globalComponents,
-      [
-        (entry) =>
-          (entry.type === "directory" && file.startsWith(entry.path)) ||
-          entry.path === file,
-      ],
+      (entry) =>
+        (entry.type === "directory" && file.startsWith(entry.path)) ||
+        entry.path === file,
     );
 
     const page = pages[0];