+++ core/source.ts
@@ -22,6 +22,7 @@ export interface Options {
   componentLoader: ComponentLoader;
   scopedData: Map<string, Data>;
   scopedPages: Map<string, Data[]>;
+  scopedComponents: Map<string, Components>;
   fs: FS;
   prettyUrls: boolean;
   components: {
@@ -60,6 +61,9 @@ export default class Source {
   /** The pages assigned per path */
   scopedPages: Map<string, Data[]>;
 
+  /** The components assigned per path */
+  scopedComponents: Map<string, Components>;
+
   /** Use pretty URLs */
   prettyUrls: boolean;
 
@@ -97,6 +101,7 @@ export default class Source {
     this.components = options.components;
     this.scopedData = options.scopedData;
     this.scopedPages = options.scopedPages;
+    this.scopedComponents = options.scopedComponents;
     this.prettyUrls = options.prettyUrls;
   }
 
@@ -118,10 +123,7 @@ export default class Source {
     );
   }
 
-  async build(
-    globalComponents: Components,
-    buildFilters: BuildFilter[],
-  ): Promise<[Page[], StaticFile[]]> {
+  async build(...buildFilters: BuildFilter[]): Promise<[Page[], StaticFile[]]> {
     const pages: Page[] = [];
     const staticFiles: StaticFile[] = [];
 
@@ -129,7 +131,7 @@ export default class Source {
       buildFilters,
       this.fs.entries.get("/")!,
       "/",
-      globalComponents,
+      new Map(),
       {},
       pages,
       staticFiles,
@@ -170,23 +172,37 @@ export default class Source {
       }
     }
 
+    // Directory data
     const dirData = mergeData(
       this.scopedData.get(dir.path) || {},
       parentData,
       currentData,
     );
 
+    // Directory components
+    const scopedComponents = this.scopedComponents.get(dir.path);
+    let loadedComponents: Components | undefined;
+
     // Load _components files
     for (const entry of dir.children.values()) {
       if (entry.type === "directory" && entry.name === "_components") {
-        parentComponents = new Map(parentComponents);
-        await this.componentLoader.load(entry, dirData, parentComponents);
-        dirData[this.components.variable] = toProxy(
-          parentComponents,
-          this.extraCode,
-        );
+        loadedComponents = await this.componentLoader.load(entry, dirData);
         break;
       }
+    }
+
+    // Merge the components
+    if (scopedComponents || loadedComponents) {
+      parentComponents = mergeComponents(
+        parentComponents,
+        scopedComponents || new Map(),
+        loadedComponents || new Map(),
+      );
+
+      dirData[this.components.variable] = toProxy(
+        parentComponents,
+        this.extraCode,
+      );
     }
 
     // Store the root data to be used by other plugins
@@ -465,27 +481,25 @@ export interface ProxyComponents {
 }
 
 /** Merge the cascade components */
-export function mergeComponents(
-  current: Components,
-  previous: Components = new Map(),
-): Components {
-  const components = new Map(previous);
+export function mergeComponents(...components: Components[]): Components {
+  return components.reduce((previous, current) => {
+    const components = new Map(previous);
 
-  for (const [key, value] of current) {
-    if (components.has(key)) {
-      const previousValue = components.get(key);
+    for (const [key, value] of current) {
+      if (components.has(key)) {
+        const previousValue = components.get(key);
 
-      if (previousValue instanceof Map && value instanceof Map) {
-        components.set(key, mergeComponents(value, previousValue));
+        if (previousValue instanceof Map && value instanceof Map) {
+          components.set(key, mergeComponents(value, previousValue));
+        } else {
+          components.set(key, value);
+        }
       } else {
         components.set(key, value);
       }
-    } else {
-      components.set(key, value);
     }
-  }
-
-  return components;
+    return components;
+  });
 }
 
 /** Merge the cascade data */